tests/requirements-test.txt
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-timeout>=2.3.0
httpx>=0.28.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=0.41b0
openinference-semantic-conventions>=0.1.0
